At the very least, we can move the configurations around
